+++ extensions/toggl-track/src/index.tsx
@@ -1,58 +1,29 @@
-import { ActionPanel, clearSearchBar, Icon, List, Action, showToast, Toast } from "@raycast/api";
+import { ActionPanel, Icon, List, Action } from "@raycast/api";
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
-import { useMemo } from "react";
 
-import { createTimeEntry, TimeEntry } from "@/api";
 import TimeEntryForm from "@/components/CreateTimeEntryForm";
 import RunningTimeEntry from "@/components/RunningTimeEntry";
 import { ExtensionContextProvider } from "@/context/ExtensionContext";
 import { formatSeconds } from "@/helpers/formatSeconds";
-import { useTimeEntries, useRunningTimeEntry } from "@/hooks";
+import { useProcessedTimeEntries } from "@/hooks/useProcessedTimeEntries";
+import { useTimeEntryActions } from "@/hooks/useTimeEntryActions";
+import { useTotalDurationToday } from "@/hooks/useTotalDurationToday";
 
 dayjs.extend(duration);
 
 function ListView() {
-  const { timeEntries, isLoadingTimeEntries, revalidateTimeEntries } = useTimeEntries();
-  const { runningTimeEntry, isLoadingRunningTimeEntry, revalidateRunningTimeEntry } = useRunningTimeEntry();
+  const {
+    timeEntries,
+    runningTimeEntry,
+    isLoading,
+    revalidateTimeEntries,
+    revalidateRunningTimeEntry,
+    timeEntriesWithUniqueProjectAndDescription,
+  } = useProcessedTimeEntries();
 
-  const isLoading = isLoadingTimeEntries || isLoadingRunningTimeEntry;
-
-  const timeEntriesWithUniqueProjectAndDescription = timeEntries.reduce<typeof timeEntries>((acc, timeEntry) => {
-    if (
-      timeEntry.id === runningTimeEntry?.id ||
-      acc.find((t) => t.description === timeEntry.description && t.project_id === timeEntry.project_id)
-    )
-      return acc;
-    return [...acc, timeEntry];
-  }, []);
-
-  const totalDurationToday = useMemo(() => {
-    let seconds = timeEntries
-      .slice(runningTimeEntry ? 1 : 0)
-      .filter((timeEntry) => dayjs(timeEntry.start).isSame(dayjs(), "day"))
-      .reduce((acc, timeEntry) => acc + timeEntry.duration, 0);
-    if (runningTimeEntry) seconds += dayjs().diff(dayjs(runningTimeEntry.start), "second");
-    return seconds;
-  }, [timeEntries, runningTimeEntry]);
-
-  async function resumeTimeEntry(timeEntry: TimeEntry) {
-    await showToast(Toast.Style.Animated, "Starting timer...");
-    try {
-      await createTimeEntry({
-        projectId: timeEntry.project_id ?? undefined,
-        workspaceId: timeEntry.workspace_id,
-        description: timeEntry.description,
-        tags: timeEntry.tags,
-        billable: timeEntry.billable,
-      });
-      revalidateRunningTimeEntry();
-      await showToast(Toast.Style.Success, "Time entry resumed");
-      await clearSearchBar({ forceScrollToTop: true });
-    } catch (e) {
-      await showToast(Toast.Style.Failure, "Failed to resume time entry");
-    }
-  }
+  const totalDurationToday = useTotalDurationToday(timeEntries, runningTimeEntry);
+  const { resumeTimeEntry } = useTimeEntryActions(revalidateRunningTimeEntry, revalidateTimeEntries);
 
   return (
     <List