+++ extensions/toggl-track/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Toggl Track Changelog
 
+## [New Feature] - 2024-07-05
+
+- Add ability to view, start, or stop existing time entries from the Menu Bar
+
 ## [Enhancements] - 2024-05-27
 
 - Prettier sort imports